Refactoring the class. Change name of the class and rename CMTimeRange timeRangeOut variable inside the MTAudioProcessingTapProcessCallback.
